fix: allow system without query signature
