+++ arepy/ecs/query/__init__.py
@@ -23,28 +23,28 @@ try:
 except ImportError:
     ...
 
-T = TypeVar("T")
-P = ParamSpec("P", bound=Type[Component])  # type: ignore
+TEntity = TypeVar("TEntity")
+TFilter = TypeVar("TFilter")
+P = ParamSpec("P")
 
 
-class EntityWith(Generic[P]): ...
+class With(Generic[P]): ...
 
 
-class EntityWithout(Generic[P]): ...
+class Without(Generic[P]): ...
 
 
-class Query(Generic[T]):
+class Query(Generic[TEntity, TFilter]):
     """
 
     Query class to filter entities based on the components they have.
 
-    TODO: test this example and fix it if needed hehe :D
     Example:
     ```python
     from arepy.ecs.query import Query, EntityWith
     from arepy.bundle.components import Transform, Rigidbody2D
 
-    def movement_system(query: Query[EntityWith[Transform, Rigidbody2D]]):
+    def movement_system(query: Query[Entity, With[Transform, Rigidbody2D]]):
         for entity in query.get_entities():
             position = entity.get_component(Transform).position
             velocity = entity.get_component(Rigidbody2D).velocity
@@ -52,16 +52,22 @@ class Query(Generic[T]):
             position.x += velocity.x
             position.y += velocity.y
     ```
+
+
+    TODO:
+        Need to improve the query system, with TEntity we need to be able to get an specific entity
+        I think we can improve the performance using only an entity with threads, so a system can get the entity and the system enqueue only the entity
+        and the threads will process every entity(in chunks) instead of processing a loop of entities.
     """
 
     def __init__(self) -> None:
         self._signature = Signature(MAX_COMPONENTS)
         self._entities: set["Entity"] = set()
-        self._kind: Union[EntityWith, EntityWithout] = None  # type: ignore
+        self._kind: Union[With, Without] = None  # type: ignore
         self._thread_id: Optional[int] = None
 
     def get_component_signature(self) -> Signature:
-        if self._kind is EntityWithout and not self._signature.was_flipped:
+        if self._kind is With and not self._signature.was_flipped:
             self._signature.flip()
 
         return self._signature
@@ -78,6 +84,8 @@ class Query(Generic[T]):
     def __iter__(self) -> Iterable["Entity"]:
         return iter(self._entities)
 
+    def fetch(self) -> TEntity: ...
+
 
 def make_query_signature(function: Callable) -> OrderedDict[Any, str]:
     """Sign the query with the components that the function needs.
@@ -90,10 +98,10 @@ def make_query_signature(function: Callable) -> OrderedDict[Any, str]:
     query_signature = get_query_from_args(func_arguments, raw_query=True)
 
     if not query_signature:
-        return OrderedDict()
+        return OrderedDict(func_arguments)
 
     raw_query: Callable[[], Query] = cast(Callable[[], Query], query_signature)
-    kind_of_result = raw_query.__args__[0]
+    kind_of_result = raw_query.__args__[1]
     required_components: tuple[Type[Component], ...] = kind_of_result.__args__[0]
     query: Query = raw_query()
     query._kind = kind_of_result
@@ -122,15 +130,16 @@ def get_annotations(function: Callable) -> OrderedDict[Any, str]:
 
 def get_query_from_args(
     args: dict[object, str], raw_query: bool = False
-) -> Callable[[], Query] | Query:
+) -> Optional[Callable[[], Query] | Query]:
 
     # When whe have a raw query(Annotation) search for it and return it
     if raw_query:
-        result = [key for key in args if key.__qualname__ == Query.__name__][0]
-        return cast(Callable[[], Query], result)
+        result = [key for key in args if key.__qualname__ == Query.__name__]
+        if not result:
+            return None
+        return cast(Callable[[], Query], result[0])
 
     # When the instance is a Query, return it ^-^
     for instance, _ in args.items():
         if isinstance(instance, Query):
             return instance
-    raise Exception("Missing Query annotation")
